Rename the `schema` directory to the more conventional `models`
