+++ src/database/models/setups.py
@@ -2,7 +2,7 @@
 
 from sqlalchemy.orm import Mapped, mapped_column
 
-from database.schema.base import Base, ExpDBReflected
+from database.models.base import Base, ExpDBReflected
 from routers.types import Identifier
