+++ src/database/studies.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, cast
 
 from sqlalchemy import text
 
-from database.schema.base import UntypedRow
+from database.models.base import UntypedRow
 from database.users import User
 from routers.types import Identifier
 from schemas.study import CreateStudy, StudyType
