+++ src/routers/openml/tasks.py
@@ -14,7 +14,7 @@ import database.tasks
 from config import get_config
 from core.errors import InternalError, NoResultsError, TagAlreadyExistsError, TaskNotFoundError
 from database.exceptions import DuplicatePrimaryKeyError, ForeignKeyConstraintError
-from database.schema.base import UntypedRow
+from database.models.base import UntypedRow
 from database.users import User
 from routers.dependencies import Pagination, expdb_connection, expdb_session, fetch_user_or_raise
 from routers.types import (
